♻️ Refactor popupContent
